requirements.txt
selenium==3.141.0
pyyaml==6.0
beautifulsoup4==4.10.0
tqdm==4.64.0
pandas==1.4.1
fake-useragent==0.1.11
pymysql==1.0.2